add all basic components of torment
This is a giant commit with the first substantial meat.  It includes the
basic components of torment for testing and a couple of small helpers.
More will be added in the next commit.

Also, there are a few skipped tests that need issues opened and a couple
of small TODO items that also need issues.