move contexts into separate module
In order to add more helpful contexts we need a place to put them.  Of
course, this requires us to increment the major number since it's a new
interface but we buy the ability to organize not just fixtures but also
contexts (the first of which will be the DockerContext).