generalize exception processing in ErrorFixture
While adding a test for the run method in ErrorFixture I found that we
don't handle generic exceptions (i.e. RuntimeError).  I've removed the
assumptive code to allow all exceptions to be handled but this means
that users will need to override ErrorFixture if they want to do custom
checking beyond the exception was raised.