add evert helper function
Evert as many are unaware is a cool word.  What this function does is
what you may have seen in futurama when Farnsworth makes a universe in a
box and ends up with a box containing the universe it's a part of.

This function can allow very complex patterns of testing composition
from an allowable set of parameters to mixin with required expectations.
Examples will need to be added to the documentation.